+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/otaupgrade/QueryNextImageCommand.java
@@ -227,12 +227,12 @@ public class QueryNextImageCommand extends ZclOtaUpgradeCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        fieldControl = (Integer) deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
-        manufacturerCode = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
-        imageType = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
-        fileVersion = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_32_BIT_INTEGER);
+        fieldControl = deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
+        manufacturerCode = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        imageType = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        fileVersion = deserializer.deserialize(ZclDataType.UNSIGNED_32_BIT_INTEGER);
         if ((fieldControl & 0x01) != 0) {
-            hardwareVersion = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+            hardwareVersion = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
         }
     }
